+++ HW5/app/src/main/java/com/kodeco/android/countryinfo/types/CountriesRepository.kt
@@ -1,22 +1,8 @@
 package com.kodeco.android.countryinfo.types
 
 import com.kodeco.android.countryinfo.ui.screens.countryList.Resource
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
-import retrofit2.Response
-import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
-import retrofit2.http.GET
-
-interface RestCountriesService {
-    @GET("all")
-    suspend fun getAllCountries(): Response<List<Country>>
-}
-
-
-
 
 interface ICountriesRepository {
     fun getAllCountries(): Flow<Resource<List<Country>>>
@@ -27,20 +13,8 @@ class CountriesRepository: ICountriesRepository {
 
     private var countries: Map<String, Country> = mutableMapOf<String, Country>()
 
-    private fun createRetrofitService(): RestCountriesService {
-        val moshi = Moshi.Builder()
-            .add(KotlinJsonAdapterFactory())
-            .build()
-
-        val retrofit = Retrofit.Builder()
-            .baseUrl("https://restcountries.com/v3.1/")
-            .addConverterFactory(MoshiConverterFactory.create(moshi))
-            .build()
-
-        return retrofit.create(RestCountriesService::class.java)
-    }
     override fun getAllCountries(): Flow<Resource<List<Country>>> = flow {
-        val service = createRetrofitService()
+        val service = CountriesService().createRetrofitService()
         try {
             emit(Resource.Loading())
             val response = service.getAllCountries()
