+++ HW5/app/src/main/java/com/kodeco/android/countryinfo/ui/components/CountryInfoRow.kt
@@ -17,14 +17,8 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
-import com.kodeco.android.countryinfo.types.CountriesRepository
 import com.kodeco.android.countryinfo.types.Country
 import com.kodeco.android.countryinfo.ui.screens.countryList.CountryInfoViewModel
-import com.kodeco.android.countryinfo.ui.screens.countryList.CountryInfoViewModelFactory
 
 @Composable
 fun CountryInfoRow(viewModel: CountryInfoViewModel, country: Country) {
@@ -47,10 +41,8 @@ fun CountryInfoRow(viewModel: CountryInfoViewModel, country: Country) {
                     Text(
                         text = country.name.common,
                         fontWeight = FontWeight.Bold,
-
                         modifier = Modifier.padding(horizontal = 16.dp, vertical = 4.dp)
                     )
-
 
                     if (country.capital?.isNotEmpty() == true) {
                         Text(
@@ -60,13 +52,6 @@ fun CountryInfoRow(viewModel: CountryInfoViewModel, country: Country) {
                     }
                 }
             }
-
-
-
-//            if (isExpanded) {
-//                CountryDetailsScreen(country = country, viewModel = viewModel)
-//            }
-
         }
 
 
@@ -79,76 +64,3 @@ fun CountryInfoRow(viewModel: CountryInfoViewModel, country: Country) {
 @Composable
 fun CountryInfoRowPreview() { }
 
-@Composable
-fun CountryInfoNavHost(repository: CountriesRepository) {
-
-    val navController = rememberNavController()
-
-//    NavHost(navController = navController, startDestination = "list")
-
-    NavHost(
-        navController = navController,
-        startDestination = "list",
-//        modifier = Modifier.padding(innerPadding)
-    ) {
-
-        composable("list") {
-
-            MainView(viewModel = viewModel(
-                factory = CountryInfoViewModelFactory(
-                    repository = repository,
-                    navController = navController
-                )
-
-            ))
-
-//            CountryInfoScreen(
-//                viewModel = viewModel(
-//                    factory = CountryInfoViewModel.CountryInfoViewModelFactory(
-//                        repository = repository,
-//                    ),
-//                ),
-//                onCountryRowTap = { countryIndex ->
-//                    navController.navigate("details/$countryIndex")
-//                }
-//            )
-        }
-
-        composable("country/{countryName}") {
-            val countryName = it.arguments?.getString("countryName") ?: error("CountryName is required")
-            val country = repository.getCountry(countryName)
-            if (country != null) {
-                CountryDetailsScreen(country = country, viewModel = viewModel(
-                    factory = CountryDetailsViewModelFactory(
-//                        repository = repository,
-                        navController = navController
-                    )
-                ))
-            }
-            else {
-                CountryErrorScreen(errorMessage = "No Country Name")
-            }
-
-//            CountryDetailsScreen(country = repository.getCountry(it.arguments.getString("countryName") ?: error("Country Name is required")), viewModel = )
-        }
-
-
-    }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-}
