+++ HW5/app/src/main/java/com/kodeco/android/countryinfo/ui/screens/countryList/CountryInfoViewModel.kt
@@ -77,4 +77,9 @@ class CountryInfoViewModel(
         navController.navigate(route = "country/$countryName")
     }
 
+
+    fun openAbout(){
+        navController.navigate("about")
+    }
+
 }
